+++ lightningd/lightningd.c
@@ -543,9 +543,8 @@ static void pidfile_create(const struct lightningd *ld)
 	int pid_fd;
 	char *pid;
 
-	/* Create PID file: relative to .config dir unless absolute. */
-	pid_fd = open(path_join(tmpctx, ld->config_dir, ld->pidfile),
-		      O_WRONLY|O_CREAT, 0640);
+	/* Create PID file: relative to .config dir. */
+	pid_fd = open(ld->pidfile, O_WRONLY|O_CREAT, 0640);
 	if (pid_fd < 0)
 		err(1, "Failed to open PID file");
 
@@ -572,11 +571,7 @@ static void pidfile_create(const struct lightningd *ld)
  * extra sanity checks, and it's also a good point to free the tmpctx. */
 static int io_poll_lightningd(struct pollfd *fds, nfds_t nfds, int timeout)
 {
-	/*~ In particular, we should *not* have left a database transaction
-	 * open! */
-	db_assert_no_outstanding_statements();
-
-	/* The other checks and freeing tmpctx are common to all daemons. */
+	/* These checks and freeing tmpctx are common to all daemons. */
 	return daemon_poll(fds, nfds, timeout);
 }